tidspbridge-extra: Create symlink to bridgedriver.ko.
The install_module script in /dspbridge requires a symlink in that directory pointing to the Bridge driver. A post-install script takes care of adding such symlink. It is important to note that it will symlink to the driver of the *running* kernel the first time the system boots. Please read that last sentence once again. Make sure you use the right kernel or maintain that symlink in working order.